DOMA-1214 fixed usage of context for ServiceSubscription
Moved `useServiceSubscription` hook locally and renamed it to `useServiceSubscriptionLoader`.
Added hook `useServiceSubscriptionContext` and replaced all usages of `useServiceSubscription` to it.
