+++ apps/condo/domains/subscription/components/SubscriptionContext.tsx
@@ -1,18 +1,55 @@
-import React, { useEffect, useState } from 'react'
+import React, { useContext, useEffect, useState, createContext } from 'react'
 import Router, { useRouter } from 'next/router'
-import { ServiceSubscription } from '../../../schema'
+import { ServiceSubscription, SortServiceSubscriptionsBy } from '../../../schema'
 import { Modal, Typography } from 'antd'
 import { useIntl } from '@core/next/intl'
 import dayjs from 'dayjs'
 import relativeTime from 'dayjs/plugin/relativeTime'
 import duration from 'dayjs/plugin/duration'
 import { Button } from '@condo/domains/common/components/Button'
-import { useServiceSubscription } from '../hooks/useServiceSubscription'
+import { useOrganization } from '../../../../../packages/@core.next/organization'
+import { useObjects } from '../utils/clientSchema/ServiceSubscription'
+import { get } from 'lodash'
+import { isExpired } from '../utils/helpers'
 
 dayjs.extend(relativeTime)
 dayjs.extend(duration)
 
-const SubscriptionContext = React.createContext({})
+interface ISubscriptionContext {
+    subscription?: ServiceSubscription
+    daysLeft?: number
+    daysLeftHumanized?: string
+    isExpired?: boolean
+}
+
+const SubscriptionContext = createContext<ISubscriptionContext>({})
+
+export const useServiceSubscriptionContext = () => useContext(SubscriptionContext)
+
+const useServiceSubscriptionLoader = (): ISubscriptionContext => {
+    const { organization } = useOrganization()
+    const { objs } = useObjects({
+        where: {
+            organization: { id: get(organization, 'id') },
+        },
+        sortBy: [SortServiceSubscriptionsBy.StartAtDesc],
+    }, {
+        fetchPolicy: 'network-only',
+    })
+    const subscription = objs[0]
+    if (!subscription) {
+        return {}
+    }
+    const daysLeftDuration = dayjs.duration(dayjs(subscription.finishAt).diff(dayjs()))
+    const daysLeft = Math.ceil(daysLeftDuration.asDays())
+    const daysLeftHumanized = daysLeftDuration.humanize()
+    return {
+        subscription,
+        daysLeft,
+        daysLeftHumanized,
+        isExpired: isExpired(subscription),
+    }
+}
 
 interface IExpiredModal {
     subscription: ServiceSubscription
@@ -52,13 +89,12 @@ const ExpiredModal: React.FC<IExpiredModal> = ({ subscription }) => {
     )
 }
 
-interface ISubscriptionProvider {
-    organizationId: number
+interface ISubscriptionProviderProps {
     children: JSX.Element
 }
 
-export const SubscriptionProvider: React.FC<ISubscriptionProvider> = ({ organizationId, children }) => {
-    const { subscription, isExpired } = useServiceSubscription()
+export const SubscriptionProvider: React.FC<ISubscriptionProviderProps> = ({ children }) => {
+    const { subscription, isExpired, daysLeft, daysLeftHumanized } = useServiceSubscriptionLoader()
     const { route } = useRouter()
     useEffect(() => {
         if (isExpired && route !== '/settings') {
@@ -69,7 +105,7 @@ export const SubscriptionProvider: React.FC<ISubscriptionProvider> = ({ organiza
         return children
     }
     return (
-        <SubscriptionContext.Provider value={subscription}>
+        <SubscriptionContext.Provider value={{ subscription, isExpired, daysLeft, daysLeftHumanized }}>
             {isExpired && (
                 <ExpiredModal subscription={subscription}/>
             )}
